requirements.txt
openai
python-dotenv
requests
PyPDF2
gradio